Refactor cache into working_directory
The cache was confusing as a concept. More clear to explicity define
usage with two methods -- one for setting the working directory, and
one for resolving paths to that working directory.

The caller can request for a path to be resolved, and then do all the
existing testing they need on their end to determine whether or not
to use a cached file. The working_directory module is now only responsible
for resolving the path if it is relative to the working directory.